Use int64 as type for queue readerID (#4176)

+++ service/history/queues/action_pending_task_count.go
@@ -45,7 +45,7 @@ type (
 	actionQueuePendingTask struct {
 		attributes     *AlertAttributesQueuePendingTaskCount
 		monitor        Monitor
-		maxReaderCount int
+		maxReaderCount int64
 
 		// state of the action, used when running the action
 		tasksPerNamespace               map[namespace.ID]int
@@ -63,7 +63,7 @@ func newQueuePendingTaskAction(
 	return &actionQueuePendingTask{
 		attributes:     attributes,
 		monitor:        monitor,
-		maxReaderCount: maxReaderCount,
+		maxReaderCount: int64(maxReaderCount),
 	}
 }
 
@@ -93,7 +93,7 @@ func (a *actionQueuePendingTask) Run(readerGroup *ReaderGroup) error {
 }
 
 func (a *actionQueuePendingTask) tryShrinkSlice(
-	readers map[int32]Reader,
+	readers map[int64]Reader,
 ) bool {
 	for _, reader := range readers {
 		reader.ShrinkSlices()
@@ -109,7 +109,7 @@ func (a *actionQueuePendingTask) init() {
 }
 
 func (a *actionQueuePendingTask) gatherStatistics(
-	readers map[int32]Reader,
+	readers map[int64]Reader,
 ) {
 	// gather statistic for
 	// 1. total # of pending tasks per namespace
@@ -176,7 +176,7 @@ func (a *actionQueuePendingTask) findSliceToClear(
 }
 
 func (a *actionQueuePendingTask) splitAndClearSlice(
-	readers map[int32]Reader,
+	readers map[int64]Reader,
 	readerGroup *ReaderGroup,
 ) error {
 	if err := a.ensureNewReaders(readers, readerGroup); err != nil {
@@ -184,7 +184,7 @@ func (a *actionQueuePendingTask) splitAndClearSlice(
 	}
 
 	for readerID, reader := range readers {
-		if readerID == int32(a.maxReaderCount)-1 {
+		if readerID == int64(a.maxReaderCount)-1 {
 			// we can't do further split, have to clear entire slice
 			cleared := false
 			reader.ClearSlices(func(s Slice) bool {
@@ -238,11 +238,11 @@ func (a *actionQueuePendingTask) splitAndClearSlice(
 }
 
 func (a *actionQueuePendingTask) ensureNewReaders(
-	readers map[int32]Reader,
+	readers map[int64]Reader,
 	readerGroup *ReaderGroup,
 ) error {
 	for readerID, reader := range readers {
-		if readerID == int32(a.maxReaderCount)-1 {
+		if readerID == a.maxReaderCount-1 {
 			// we won't perform split
 			continue
 		}
